# api/auth/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    name: AussieBroadWAN Team
    url: https://github.com/aussiebroadwan/bartab
  description: |-
    Attempting to be OAuth2-compliant for our authentication service providing token management with JWT-based access tokens.

    All tokens are signed using RS256 (RSA-SHA256) and can be verified using the JWKS endpoint.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: BarTab Authentication Service API
  version: 0.1.0
paths:
...
  /v1/oauth2/authorize:
    get:
      description: |-
        Initiates the OAuth2 authorization code flow via GET request. Used for browser redirects.
        If a valid session exists (cookie or Bearer token), issues authorization code immediately.
        Otherwise, returns 401 with login_required error.

        **PKCE Support:**
        - Public clients MUST include code_challenge (defaults to S256 if method omitted)
        - Confidential clients MAY include code_challenge for additional security

        **Response:**
        - Success: 302 redirect to redirect_uri with code and state parameters
        - No session: 401 JSON with login_required error
        - Error: JSON error response
      parameters:
      - default: code
        description: Must be 'code'
        in: query
        name: response_type
        required: true
        type: string
      - description: OAuth2 client identifier
        in: query
        name: client_id
        required: true
        type: string
      - description: Callback URI (must match registered redirect URI)
        in: query
        name: redirect_uri
        required: true
        type: string
      - description: Space-delimited list of scopes
        example: '"profile:read admin:write"'
        in: query
        name: scope
        type: string
      - description: Opaque value for CSRF protection (recommended)
        in: query
        name: state
        type: string
      - description: PKCE code challenge (required for public clients)
        example: '"E9Melhoa2OwvFrEMTJguCHaoeK1t8URWbuGJSstw-cM"'
        in: query
        name: code_challenge
        type: string
      - default: S256
        description: PKCE method (S256 or plain, defaults to S256)
        enum:
        - S256
        - plain
        in: query
        name: code_challenge_method
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to redirect_uri with code and state
          schema:
            type: string
        "400":
          description: "Invalid request\"\texample({\"error\":\"invalid_request\",\"error_description\":\"missing
            required parameter\"})"
          schema:
            additionalProperties: true
            type: object
        "401":
          description: "Unauthorized\"\t\texample({\"error\":\"login_required\",\"error_description\":\"user
            authentication required\"})"
          schema:
            additionalProperties: true
            type: object
      summary: OAuth2 authorization endpoint (GET)
      tags:
      - OAuth2
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: |-
        Initiates the OAuth2 authorization code flow via POST request with credentials.
        Supports three authentication methods:
        - Session-based: Include session cookie or Bearer token in Authorization header
        - Username/Password: Submit credentials in POST body
        - MFA Completion: Submit mfa_token with mfa_method and mfa_code

        **PKCE Support:**
        - Public clients MUST include code_challenge (defaults to S256 if method omitted)
        - Confidential clients MAY include code_challenge for additional security

        **Response:**
        - Success: 302 redirect to redirect_uri with code and state parameters
        - MFA Required: 409 JSON response with mfa_token and available methods
        - Error: Either redirect with error parameters or JSON error response
      parameters:
      - default: code
        description: Must be 'code'
        in: query
        name: response_type
        required: true
        type: string
      - description: OAuth2 client identifier
        in: query
        name: client_id
        required: true
        type: string
      - description: Callback URI (must match registered redirect URI)
        in: query
        name: redirect_uri
        required: true
        type: string
      - description: Space-delimited list of scopes
        example: '"profile:read admin:write"'
        in: query
        name: scope
        type: string
      - description: Opaque value for CSRF protection (recommended)
        in: query
        name: state
        type: string
      - description: PKCE code challenge (required for public clients)
        example: '"E9Melhoa2OwvFrEMTJguCHaoeK1t8URWbuGJSstw-cM"'
        in: query
        name: code_challenge
        type: string
      - default: S256
        description: PKCE method (S256 or plain, defaults to S256)
        enum:
        - S256
        - plain
        in: query
        name: code_challenge_method
        type: string
      - description: Username for password authentication
        in: formData
        name: username
        type: string
      - description: Password for password authentication
        in: formData
        name: password
        type: string
      - description: MFA token from previous 409 response
        in: formData
        name: mfa_token
        type: string
      - description: MFA method (totp or backup_codes)
        enum:
        - totp
        - backup_codes
        in: formData
        name: mfa_method
        type: string
      - description: MFA code (6-digit TOTP or backup code)
        in: formData
        name: mfa_code
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to redirect_uri with code and state
          schema:
            type: string
        "400":
          description: "Invalid request\"\texample({\"error\":\"invalid_request\",\"error_description\":\"missing
            required parameter\"})"
          schema:
            additionalProperties: true
            type: object
        "401":
          description: "Unauthorized\"\t\texample({\"error\":\"login_required\",\"error_description\":\"user
            authentication required\"})"
          schema:
            additionalProperties: true
            type: object
        "409":
          description: "MFA required\"\t\texample({\"error\":\"mfa_required\",\"error_description\":\"multi-factor
            authentication is required\",\"mfa_token\":\"token123\",\"mfa_methods\":[\"totp\",\"backup_codes\"]})"
          schema:
            additionalProperties: true
            type: object
      summary: OAuth2 authorization endpoint (POST)
      tags:
      - OAuth2
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: "Bearer {token}".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"